original_boilerplate/requirements.txt
yfinance==0.1.55
pandas==1.1.5
numpy==1.19.4 
statsmodels==0.12.1 
pyotp==2.4.1 
robin-stocks==1.5.3
matplotlib==3.3.3
alpaca-trade-api==0.51.0 
python-telegram-bot==12.2.0